move hidden activities to activity module
