+++ packages/shared/src/lib/core/activities/stores/index.ts
@@ -1,3 +1,4 @@
 export * from './claimed-activities.store'
+export * from './hidden-activities.store'
 export * from './persisted-balance-changes.store'
 export * from './persisted-transactions.store'
