+++ packages/shared/src/lib/core/activities/types/hidden-activities.interface.ts
