+++ packages/shared/src/lib/core/activities/types/index.ts
@@ -1,3 +1,4 @@
 export * from './asset-balance-change.interface'
 export * from './claimed-activities.interface'
+export * from './hidden-activities.interface'
 export * from './persisted-evm-transaction.interface'
