+++ packages/shared/src/lib/core/wallet/actions/rejectActivity.ts
@@ -1,7 +1,7 @@
 import { selectedAccount } from '@core/account'
 import { get } from 'svelte/store'
 import { updateAsyncDataByActivityId } from '../stores'
-import { hiddenActivities } from '../stores/hidden-activities.store'
+import { hiddenActivities } from '../../activities/stores/hidden-activities.store'
 import { localize } from '@core/i18n'
 import { showAppNotification } from '@auxiliary/notification'
 import { activeProfileId } from '@core/profile'
