+++ packages/shared/src/lib/core/wallet/interfaces/index.ts
@@ -9,7 +9,6 @@ export * from './asset.interface'
 export * from './base-token.interface'
 export * from './contact-subject.interface'
 export * from './erc20-metadata.interface'
-export * from './hidden-activities.interface'
 export * from './irc30-metadata.interface'
 export * from './mint-nft-details.interface'
 export * from './mint-token-details.interface'
